FreetrailPoseEstimation/requirements.txt
opencv-python
mediapipe
numpy
flask_cors